:slot_machine: Extract generator function into its own file to allow testing
